packages/eap-core/pyproject.toml
[project]
name = "eap-core"
version = "0.1.0"
description = "Enterprise Agentic AI Platform SDK — core middleware, runtimes, identity"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "EAP-Core Authors" }]
dependencies = [
    "pydantic>=2.6",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.8",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
]

[project.optional-dependencies]
pii = ["presidio-analyzer>=2.2", "presidio-anonymizer>=2.2", "spacy>=3.7"]
otel = ["opentelemetry-api>=1.24", "opentelemetry-sdk>=1.24", "opentelemetry-exporter-otlp>=1.24"]
aws = ["boto3>=1.34"]
gcp = ["google-cloud-aiplatform>=1.50"]
mcp = ["mcp>=0.9"]
a2a = ["fastapi>=0.110", "uvicorn>=0.29"]
eval = ["ragas>=0.1", "datasets>=2.18"]
policy-cedar = ["cedarpy>=2.4"]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "pytest-cov>=5", "ruff>=0.4", "mypy>=1.10"]

[project.entry-points."eap_core.runtimes"]
local = "eap_core.runtimes.local:LocalRuntimeAdapter"
bedrock = "eap_core.runtimes.bedrock:BedrockRuntimeAdapter"
vertex = "eap_core.runtimes.vertex:VertexRuntimeAdapter"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eap_core"]
